update transformer model
